@@ -1,91 +1,48 @@
 {
-  "name": "tacodelite-app-d53a",
+  "name": "tacodelite-app",
   "private": true,
-  "sideEffects": false,
+  "version": "2.0.0",
+  "type": "module",
   "scripts": {
-    "build": "run-s build:*",
-    "build:css": "npm run generate:css -- --minify",
-    "build:remix": "remix build",
-    "dev": "run-p dev:*",
-    "dev:css": "npm run generate:css -- --watch",
-    "dev:remix": "cross-env NODE_ENV=development binode --require ./mocks -- @remix-run/dev:remix dev",
-    "format": "prettier --write .",
-    "generate:css": "tailwindcss -i ./app/styles/app.css -o ./app/styles/tailwind.css",
-    "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
-    "start": "remix-serve build",
-    "start:mocks": "binode --require ./mocks -- @remix-run/serve:remix-serve build",
-    "test": "vitest",
-    "test:e2e:dev": "start-server-and-test dev http://localhost:3000 \"npx cypress open\"",
-    "pretest:e2e:run": "npm run build",
-    "test:e2e:run": "cross-env PORT=8811 start-server-and-test start:mocks http://localhost:8811 \"npx cypress run\"",
-    "typecheck": "tsc && tsc -p cypress",
-    "validate": "run-p \"test -- --run\" lint typecheck test:e2e:run",
-    "ts-node": "ts-node --compiler-options '{\"module\":\"CommonJS\"}'"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview",
+    "deploy": "npm run build && npm run deploy:terraform",
+    "deploy:terraform": "terraform -chdir=. init && terraform -chdir=. plan -var-file=staging.tfvars && terraform -chdir=. apply -var-file=staging.tfvars -auto-approve",
+    "deploy:staging": "npm run build && terraform -chdir=. init && terraform -chdir=. plan -var-file=staging.tfvars && terraform -chdir=. apply -var-file=staging.tfvars -auto-approve",
+    "deploy:production": "npm run build && terraform -chdir=. plan -var-file=production.tfvars && terraform -chdir=. apply -var-file=production.tfvars -auto-approve",
+    "terraform:init": "terraform -chdir=. init",
+    "terraform:plan": "terraform -chdir=. plan -var-file=staging.tfvars",
+    "terraform:apply": "terraform -chdir=. apply -var-file=staging.tfvars -auto-approve",
+    "terraform:destroy": "terraform -chdir=. destroy -var-file=staging.tfvars -auto-approve",
+    "terraform:output": "terraform -chdir=. output",
+    "upload:static": "npm run build && aws s3 sync dist/ s3://$(terraform -chdir=. output -raw s3_bucket_name) --delete",
+    "seed:db": "node seed-dynamodb.js"
   },
-  "prettier": {},
-  "eslintIgnore": [
-    "/node_modules",
-    "/build",
-    "/public/build"
-  ],
   "dependencies": {
-    "@remix-run/node": "^1.13.0",
-    "@remix-run/react": "^1.13.0",
-    "@remix-run/serve": "^1.13.0",
-    "@remix-run/server-runtime": "^1.13.0",
-    "@supabase/auth-helpers-remix": "^0.2.2",
     "@supabase/supabase-js": "^2.38.2",
-    "bcryptjs": "^2.4.3",
     "framer-motion": "^9.0.3",
-    "isbot": "^3.6.5",
     "react": "^18.2.0",
     "react-datepicker": "^4.10.0",
     "react-dom": "^18.2.0",
     "react-intersection-observer": "^9.4.2",
-    "tiny-invariant": "^1.3.1"
+    "react-router-dom": "^7.0.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^7.6.0",
-    "@remix-run/dev": "^1.13.0",
-    "@remix-run/eslint-config": "^1.13.0",
-    "@testing-library/cypress": "^8.0.7",
-    "@testing-library/dom": "^8.20.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/bcryptjs": "^2.4.2",
-    "@types/eslint": "^8.4.10",
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.10",
-    "@vitejs/plugin-react": "^3.0.1",
-    "@vitest/coverage-c8": "^0.27.2",
-    "autoprefixer": "^10.4.13",
-    "binode": "^1.0.5",
-    "c8": "^7.12.0",
-    "cookie": "^0.5.0",
-    "cross-env": "^7.0.3",
-    "cypress": "^10.11.0",
-    "eslint": "^8.32.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "happy-dom": "^8.1.4",
-    "msw": "^0.49.2",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.21",
-    "prettier": "2.8.3",
-    "prettier-plugin-tailwindcss": "^0.2.1",
-    "prisma": "^5.5.2",
-    "start-server-and-test": "^1.15.2",
-    "tailwindcss": "^3.2.4",
-    "ts-node": "^10.9.1",
-    "tsconfig-paths": "^4.1.2",
-    "typescript": "^4.9.4",
-    "vite": "^4.0.4",
-    "vite-tsconfig-paths": "^3.6.0",
-    "vitest": "^0.27.2"
-  },
-  "engines": {
-    "node": ">=14"
+    "@aws-sdk/client-dynamodb": "^3.0.0",
+    "@aws-sdk/lib-dynamodb": "^3.0.0",
+    "@eslint/js": "^9.33.0",
+    "@types/cypress": "^0.1.6",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.16",
+    "csv-parser": "^3.0.0",
+    "eslint": "^8.55.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.4.0",
+    "vite": "^5.0.8"
   }
 }
