@@ -1,95 +1,68 @@
-import { faker } from "@faker-js/faker";
+// ***********************************************
+// This example commands.ts shows you how to
+// create various custom commands and overwrite
+// existing commands.
+//
+// For more comprehensive examples of custom
+// commands please read more here:
+// https://on.cypress.io/custom-commands
+// ***********************************************
 
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      /**
-       * Logs in with a random user. Yields the user and adds an alias to the user
-       *
-       * @returns {typeof login}
-       * @memberof Chainable
-       * @example
-       *    cy.login()
-       * @example
-       *    cy.login({ email: 'whatever@example.com' })
-       */
-      login: typeof login;
+// -- This is a parent command --
+// Cypress.Commands.add('login', (email, password) => { ... })
+//
+//
+// -- This is a child command --
+// Cypress.Commands.add('drag', { prevSubject: 'element'}, (subject, options) => { ... })
+//
+//
+// -- This is a dual command --
+// Cypress.Commands.add('dismiss', { prevSubject: 'optional'}, (subject, options) => { ... })
+//
+//
+// -- This will overwrite an existing command --
+// Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
+//
+// declare global {
+//   namespace Cypress {
+//     interface Chainable {
+//       login(email: string, password: string): Chainable<void>
+//       drag(subject: string, options?: Partial<TypeOptions>): Chainable<Element>
+//       dismiss(subject: string, options?: Partial<TypeOptions>): Chainable<Element>
+//       visit(originalFn: CommandOriginalFn, url: string, options: Partial<VisitOptions>): Chainable<Element>
+//     }
+//   }
+// }
 
-      /**
-       * Deletes the current @user
-       *
-       * @returns {typeof cleanupUser}
-       * @memberof Chainable
-       * @example
-       *    cy.cleanupUser()
-       * @example
-       *    cy.cleanupUser({ email: 'whatever@example.com' })
-       */
-      cleanupUser: typeof cleanupUser;
+// Custom commands for Taco Delite App
 
-      /**
-       * Extends the standard visit command to wait for the page to load
-       *
-       * @returns {typeof visitAndCheck}
-       * @memberof Chainable
-       * @example
-       *    cy.visitAndCheck('/')
-       *  @example
-       *    cy.visitAndCheck('/', 500)
-       */
-      visitAndCheck: typeof visitAndCheck;
-    }
-  }
-}
+Cypress.Commands.add('login', (email: string, password: string) => {
+  // Implementation for user login
+  // This would typically interact with your auth service
+  cy.log(`Logging in user: ${email}`)
 
-function login({
-  email = faker.internet.email(undefined, undefined, "example.com"),
-}: {
-  email?: string;
-} = {}) {
-  cy.then(() => ({ email })).as("user");
-  cy.exec(
-    `npx ts-node --require tsconfig-paths/register ./cypress/support/create-user.ts "${email}"`
-  ).then(({ stdout }) => {
-    const cookieValue = stdout
-      .replace(/.*<cookie>(?<cookieValue>.*)<\/cookie>.*/s, "$<cookieValue>")
-      .trim();
-    cy.setCookie("__session", cookieValue);
-  });
-  return cy.get("@user");
-}
+  // Add your login logic here
+  // For example, if using Supabase:
+  // cy.visit('/login')
+  // cy.get('[data-testid="email-input"]').type(email)
+  // cy.get('[data-testid="password-input"]').type(password)
+  // cy.get('[data-testid="login-button"]').click()
+})
 
-function cleanupUser({ email }: { email?: string } = {}) {
-  if (email) {
-    deleteUserByEmail(email);
-  } else {
-    cy.get("@user").then((user) => {
-      const email = (user as { email?: string }).email;
-      if (email) {
-        deleteUserByEmail(email);
-      }
-    });
-  }
-  cy.clearCookie("__session");
-}
+Cypress.Commands.add('logout', () => {
+  // Implementation for user logout
+  cy.log('Logging out user')
 
-function deleteUserByEmail(email: string) {
-  cy.exec(
-    `npx ts-node --require tsconfig-paths/register ./cypress/support/delete-user.ts "${email}"`
-  );
-  cy.clearCookie("__session");
-}
+  // Add your logout logic here
+  // For example:
+  // cy.get('[data-testid="logout-button"]').click()
+  // cy.clearLocalStorage()
+  // cy.clearCookies()
+})
 
-// We're waiting a second because of this issue happen randomly
-// https://github.com/cypress-io/cypress/issues/7306
-// Also added custom types to avoid getting detached
-// https://github.com/cypress-io/cypress/issues/7306#issuecomment-1152752612
-// ===========================================================
-function visitAndCheck(url: string, waitTime: number = 1000) {
-  cy.visit(url);
-  cy.location("pathname").should("contain", url).wait(waitTime);
-}
-
-Cypress.Commands.add("login", login);
-Cypress.Commands.add("cleanupUser", cleanupUser);
-Cypress.Commands.add("visitAndCheck", visitAndCheck);
+Cypress.Commands.add('clearAuth', () => {
+  // Clear all authentication data
+  cy.clearLocalStorage()
+  cy.clearCookies()
+  cy.log('Authentication data cleared')
+})
